+++ .gitignore
@@ -21,4 +21,3 @@ yarn-error.log*
 
 #TODO add .env to your gitignore
 server/.env
-client/.env
