+++ client/src/App.js
@@ -1,30 +1,32 @@
 import React, { useEffect, useReducer } from "react";
 import { MuiThemeProvider } from "@material-ui/core";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
+import { BrowserRouter, Route, Redirect } from "react-router-dom";
+
 import { AuthStateContext, AuthDispatchContext } from "./context/AuthContext";
 import { getUser } from "./actions/auth";
 import { initialState, AuthReducer } from "./reducers/auth";
+
 import PrivateRoute from "./routing/PrivateRoute";
+
 import { theme } from "./themes/theme";
 import Signup from "./pages/Signup";
 import Login from "./pages/Login";
+import TestingRoute from "./pages/TestingRoute";
 import Navbar from "./components/navbar/Navbar";
 import Profile from "./pages/Profile";
-import EditProfile from "./pages/EditProfile";
 import Photo from "./pages/Photo";
 import Availability from "./pages/Availability";
 import Payment from "./pages/Payment";
 import Security from "./pages/Security";
 import Settings from "./pages/Settings";
-
 import ProtectedRoute from "./routes/ProtectedRoute";
 import ProfileListings from "./pages/ProfileListings";
 
 //import "./App.css";
-import PageNotFound from "./pages/PageNotFound";
 
 function App() {
   const [state, dispatch] = useReducer(AuthReducer, initialState);
+
   // Check if user is logged in when App mounts.
   useEffect(() => {
     getUser(dispatch);
@@ -36,30 +38,19 @@ function App() {
         <AuthStateContext.Provider value={state}>
           <AuthDispatchContext.Provider value={dispatch}>
             <Navbar />
-            <Switch>
-              <Route path='/signup' component={Signup} />
-              <Route path='/login' component={Login} />
-              <PrivateRoute path='/' exact component={Profile} />
-              <PrivateRoute
-                path='/dashboard/profile'
-                exact
-                component={Profile}
-              />
-              <PrivateRoute
-                path='/dashboard/editprofile'
-                component={EditProfile}
-              />
-              <PrivateRoute path='/dashboard/photo' component={Photo} />
-              <PrivateRoute
-                path='/dashboard/availability'
-                component={Availability}
-              />
-              <PrivateRoute path='/dashboard/payment' component={Payment} />
-              <PrivateRoute path='/dashboard/security' component={Security} />
-              <PrivateRoute path='/dashboard/settings' component={Settings} />
-              <PrivateRoute path='*' component={PageNotFound} />
-              <PrivateRoute path='/profile/lists' component={ProfileListings} />
-            </Switch>
+            <Route path='/signup' component={Signup} />
+            <Route path='/login' component={Login} />
+            <PrivateRoute path='/profile/lists' component={ProfileListings} />
+            <PrivateRoute path='/' exact component={Profile} />
+            <PrivateRoute path='/dashboard/profile' component={Profile} />
+            <PrivateRoute path='/dashboard/photo' component={Photo} />
+            <PrivateRoute
+              path='/dashboard/availability'
+              component={Availability}
+            />
+            <PrivateRoute path='/dashboard/payment' component={Payment} />
+            <PrivateRoute path='/dashboard/security' component={Security} />
+            <PrivateRoute path='/dashboard/settings' component={Settings} />
           </AuthDispatchContext.Provider>
         </AuthStateContext.Provider>
       </BrowserRouter>
