+++ client/src/actions/auth.js
@@ -8,7 +8,7 @@ import {
 } from "./types";
 
 // Load User (If User Logged In)
-export const getUser = async (dispatch) => {
+export const getUser = async dispatch => {
   try {
     const res = await axios.get("/api/user/getuser");
     dispatch({ type: LOGGED_IN, payload: res.data });
@@ -30,7 +30,7 @@ export const register = async (dispatch, payload) => {
     const body = JSON.stringify({ firstName, lastName, email, password });
     const res = await axios.post("/api/user/register", body, config);
     if (res.data) {
-      dispatch({ type: REGISTER_SUCCESS, payload: res.data.data });
+      dispatch({ type: REGISTER_SUCCESS, payload: res.data });
       return res.data;
     }
     // dispatch({ type: "REGISTER_ERROR", error: "a validation error ocurred" }); BACKEND ERRORS - Still to implement
@@ -52,7 +52,7 @@ export const login = async (dispatch, payload) => {
     const body = JSON.stringify({ email, password });
     const res = await axios.post("/api/user/login", body, config);
     if (res.data) {
-      dispatch({ type: LOGIN_SUCCESS, payload: res.data.data });
+      dispatch({ type: LOGIN_SUCCESS, payload: res.data });
       // dispatch({ type: "LOGIN_ERROR", error: "an error ocurred" }); BACKEND ERRORS - Still to implement
     }
   } catch (err) {
@@ -61,7 +61,7 @@ export const login = async (dispatch, payload) => {
 };
 
 // Logout User
-export const logout = async (dispatch) => {
+export const logout = async dispatch => {
   try {
     await axios.get("/api/user/logout");
     dispatch({ type: LOGOUT_SUCCESS });
