+++ client/src/components/LoginForm.js
@@ -1,6 +1,7 @@
 import React, { useState, useContext } from "react";
-import { Redirect, useLocation } from "react-router-dom";
+import { Redirect } from "react-router-dom";
 import { Typography, Grid, TextField, Button } from "@material-ui/core";
+//import { withStyles } from "@material-ui/core/styles";
 import { makeStyles } from "@material-ui/core";
 import { Link } from "react-router-dom";
 import AlertMessage from "./Alert";
@@ -8,7 +9,7 @@ import AlertMessage from "./Alert";
 import { login } from "../actions/auth";
 import { AuthDispatchContext, AuthStateContext } from "../context/AuthContext";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles(theme => ({
   root: {
     flexGrow: 1,
   },
@@ -34,16 +35,12 @@ function LoginForm() {
   const dispatch = useContext(AuthDispatchContext);
   const { isAuthenticated } = useContext(AuthStateContext);
 
-
-  //redirect to where user comes from after authentication
-  const { state } = useLocation();
-
   const handleInputChange = e => {
     setCredentials({ ...credentials, [e.target.id]: e.target.value });
     setAlert({ error: false, message: "" });
   };
   //submitting user's credentials
-  const handleSubmit = (e) => {
+  const handleSubmit = e => {
     e.preventDefault();
     //validating user input fields before submit
     if (credentials.email.length < 1 || !credentials.email) {
@@ -69,8 +66,7 @@ function LoginForm() {
 
   // Redirect if logged in
   if (isAuthenticated) {
-    return <Redirect to={state?.from || "/"} />;
-
+    return <Redirect to="/dashboard/profile" />;
   }
 
   return (
