+++ client/src/components/SignupForm.js
@@ -10,7 +10,7 @@ import { register } from "../actions/auth";
 import { AuthDispatchContext, AuthStateContext } from "../context/AuthContext";
 //import { UserContext } from "../context/Context";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles(theme => ({
   root: {
     flexGrow: 1,
   },
@@ -34,7 +34,7 @@ function SignupForm() {
   //state for alert message to pass into Alert.js component if form validation fails
   const [alert, setAlert] = useState({ error: false, message: "" });
 
-  const handleInputChange = (e) => {
+  const handleInputChange = e => {
     setCredentials({ ...credentials, [e.target.id]: e.target.value });
     setAlert({ error: false, message: "" });
   };
@@ -46,7 +46,7 @@ function SignupForm() {
   const { isAuthenticated } = useContext(AuthStateContext);
 
   //submitting user's credentials
-  const handleSubmit = (e) => {
+  const handleSubmit = e => {
     e.preventDefault();
     //validating user input fields before submit
     if (credentials.email.length < 1 || !credentials.email) {
@@ -92,7 +92,7 @@ function SignupForm() {
 
   // Redirect if logged in
   if (isAuthenticated) {
-    return <Redirect to="/dashboard/editprofile" />;
+    return <Redirect to="/dashboard/profile" />;
   }
 
   return (
