+++ client/src/components/navbar/Navbar.js
@@ -6,17 +6,14 @@ import {
   Button,
   Box,
   Hidden,
-  Avatar,
 } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core";
 import MenuIcon from "@material-ui/icons/Menu";
-import MailIcon from "@material-ui/icons/Mail";
 import { Link } from "react-router-dom";
 import MobileNavbar from "./mobileNavbar";
 import { AuthStateContext } from "../../context/AuthContext";
 import { UserContext } from "../../context/Context";
 import logo from "../../img/logo.png";
-import defaultPicture from "../../img/profile-default.png";
 
 const useStyles = makeStyles(theme => ({
   root: { flexgrow: 1 },
@@ -27,7 +24,6 @@ const useStyles = makeStyles(theme => ({
   },
   toolbarAuth: {
     backgroundColor: "white",
-    padding: "5px",
   },
   sitterLink: { color: "black", marginRight: "30px" },
   logo: {
@@ -38,8 +34,7 @@ const useStyles = makeStyles(theme => ({
 }));
 
 function Navbar() {
-  const { setMobileMenuOpen } = useContext(UserContext);
-
+  const { mobileMenuOpen, setMobileMenuOpen } = useContext(UserContext);
   const { isAuthenticated } = useContext(AuthStateContext);
   const classes = useStyles();
 
@@ -54,62 +49,32 @@ function Navbar() {
             edge="start"
             className={classes.menuButton}
             onClick={() => setMobileMenuOpen(true)}
+            style={{ marginRight: "auto" }}
           >
             <MenuIcon />
           </IconButton>
-          <MobileNavbar />
+          <MobileNavbar open={mobileMenuOpen} />
         </Hidden>
         <img src={logo} alt="logo" className={classes.logo} />
-        {isAuthenticated ? (
-          <>
-            <Hidden smDown>
-              <Link
-                to="*"
-                style={{ textDecoration: "none", marginRight: "30px" }}
-              >
-                <Button size="large">Messages</Button>
-              </Link>
-            </Hidden>
-            <Hidden mdUp>
-              <Link
-                to="*"
-                style={{ textDecoration: "none", marginRight: "30px" }}
-              >
-                <MailIcon color="primary" fontSize="large" />
-              </Link>
-            </Hidden>
-            <Link to="/dashboard/profile" style={{ textDecoration: "none" }}>
-              <Avatar
-                alt="user"
-                src={
-                  profile && profile.profilePicture
-                    ? profile.profilePicture
-                    : defaultPicture
-                }
-              />
-            </Link>
-          </>
-        ) : (
-          <Hidden smDown>
-            <Box mr={4} ml={4}>
-              <Link to="*" style={{ textDecorationColor: "black" }}>
-                <Button size="large" className={classes.sitterLink}>
-                  Become a Sitter
-                </Button>
-              </Link>
-              <Link to="/login" style={{ textDecoration: "none" }}>
-                <Button color="primary" variant="outlined" size="large">
-                  Login
-                </Button>
-              </Link>
-            </Box>
-            <Link to="/signup" style={{ textDecoration: "none" }}>
-              <Button color="primary" variant="contained" size="large">
-                Sign Up
+        <Hidden smDown>
+          <Box mr={4} ml={4}>
+            <Link to="*" style={{ textDecorationColor: "black" }}>
+              <Button size="large" className={classes.sitterLink}>
+                Become a Sitter
               </Button>
             </Link>
-          </Hidden>
-        )}
+            <Link to="/login" style={{ textDecoration: "none" }}>
+              <Button color="primary" variant="outlined" size="large">
+                Login
+              </Button>
+            </Link>
+          </Box>
+          <Link to="/signup" style={{ textDecoration: "none" }}>
+            <Button color="primary" variant="contained" size="large">
+              Sign Up
+            </Button>
+          </Link>
+        </Hidden>
       </Toolbar>
     </AppBar>
   );
