+++ client/src/reducers/auth.js
@@ -31,8 +31,8 @@ export const AuthReducer = (state = initialState, action) => {
       return {
         ...state,
         isAuthenticated: true,
-        user: payload.user,
-        profile: payload.user.profile,
+        user: payload.data,
+        profile: payload.data.profile,
         loading: false,
       };
     case PROFILE_UPDATE_SUCCESS:
