--- server/models/requestModel.js
@@ -1,40 +1,0 @@
-const mongoose = require("mongoose");
-
-const requestSchema = new mongoose.Schema({
-  ownerId: {
-    type: Schema.Types.ObjectId,
-    ref: "Profile",
-    index: true,
-    required: true,
-  },
-  sitterId: {
-    type: Schema.Types.ObjectId,
-    ref: "Profile",
-    index: true,
-    required: true,
-  },
-  start: {
-    type: Date,
-    required: true,
-  },
-  end: {
-    type: Date,
-    required: true,
-  },
-  accepted: {
-    type: Boolean,
-    default: false,
-  },
-  declined: {
-    type: Boolean,
-    default: false,
-  },
-  paid: {
-    type: Boolean,
-    default: false,
-  },
-});
-
-const Request = mongoose.model("Request", requestSchema);
-
-module.exports = Request;
