+++ server/routes/profileRouter.js
@@ -11,6 +11,5 @@ router.put("/:id", profileController.updateProfile); // update /profile:id
 router.get("/:id", profileController.getProfile); // get /profile:id
 router.post("/upload/:id", profileController.upload); // upload profile picture
 router.get("/list/:id", profileController.getProfileList); // get /profile -- returns a list of profiles
-router.delete("/delete/:id", profileController.deletePicture); // delete profile picture
 
 module.exports = router;
